move TxExecutionInfo from rpc_utils to utils
