+++ crates/bin/prove_block/src/utils.rs
@@ -1,7 +1,12 @@
-use std::collections::HashSet;
+use std::collections::{HashMap, HashSet};
 
+use blockifier::execution::call_info::CallInfo;
+use blockifier::transaction::objects::TransactionExecutionInfo;
 use cairo_vm::Felt252;
 use starknet::core::types::{ExecuteInvocation, FunctionInvocation, TransactionTrace, TransactionTraceWithHash};
+use starknet_api::core::ContractAddress;
+use starknet_api::state::StorageKey;
+use starknet_os::utils::felt_api2vm;
 
 /// Receives the transaction traces of a given block
 /// And extract the contracts addresses that where subcalled
@@ -40,4 +45,67 @@ fn process_function_invocations(inv: &FunctionInvocation, contracts: &mut HashSe
     for call in &inv.calls {
         process_function_invocations(call, contracts);
     }
+}
+
+/// Utility to get all the accesed keys from TxexecutionInfo resulted from
+/// Reexecuting all block tx using blockifier
+/// We need this as the OS require proofs for all the accessed values
+pub(crate) fn get_all_accessed_keys(
+    tx_execution_infos: &[TransactionExecutionInfo],
+) -> HashMap<ContractAddress, HashSet<StorageKey>> {
+    let mut accessed_keys_by_address: HashMap<ContractAddress, HashSet<StorageKey>> = HashMap::new();
+
+    for tx_execution_info in tx_execution_infos {
+        let accessed_keys_in_tx = get_accessed_keys_in_tx(tx_execution_info);
+        for (contract_address, storage_keys) in accessed_keys_in_tx {
+            accessed_keys_by_address.entry(contract_address).or_default().extend(storage_keys);
+        }
+    }
+
+    accessed_keys_by_address
+}
+
+fn get_accessed_keys_in_tx(
+    tx_execution_info: &TransactionExecutionInfo,
+) -> HashMap<ContractAddress, HashSet<StorageKey>> {
+    let mut accessed_keys_by_address: HashMap<ContractAddress, HashSet<StorageKey>> = HashMap::new();
+
+    for call_info in [
+        &tx_execution_info.validate_call_info,
+        &tx_execution_info.execute_call_info,
+        &tx_execution_info.fee_transfer_call_info,
+    ]
+    .into_iter()
+    .flatten()
+    {
+        let call_storage_keys = get_accessed_storage_keys(call_info);
+        for (contract_address, storage_keys) in call_storage_keys {
+            accessed_keys_by_address.entry(contract_address).or_default().extend(storage_keys);
+        }
+    }
+
+    accessed_keys_by_address
+}
+
+fn get_accessed_storage_keys(call_info: &CallInfo) -> HashMap<ContractAddress, HashSet<StorageKey>> {
+    let mut accessed_keys_by_address: HashMap<ContractAddress, HashSet<StorageKey>> = HashMap::new();
+
+    let contract_address = &call_info.call.storage_address;
+    accessed_keys_by_address
+        .entry(*contract_address)
+        .or_default()
+        .extend(call_info.accessed_storage_keys.iter().copied());
+
+    let storage_keys: Vec<_> =
+        call_info.accessed_storage_keys.iter().map(|x| felt_api2vm(*x.key()).to_hex_string()).collect();
+    log::debug!("{}: {:?}", contract_address.to_string(), storage_keys);
+
+    for inner_call in &call_info.inner_calls {
+        let inner_call_storage_keys = get_accessed_storage_keys(inner_call);
+        for (contract_address, storage_keys) in inner_call_storage_keys {
+            accessed_keys_by_address.entry(contract_address).or_default().extend(storage_keys);
+        }
+    }
+
+    accessed_keys_by_address
 }
